Improve keybinding descriptions
This commit does two things:
1) Adds tooltips to a lot of keybindings
2) Selects keybindings to display in the options view

We now have three different string fields associated with a keybinding: the description, the short
description, and the tooltip. The short description, if it is defined, will appear in the options view,
assuming Display is true, otherwise we fallback to the regular description.

I will admit, when I started this work, I didn't have a 'short' description: it was just the description
and the tooltip, and I figured I wanted to shorted the descriptions and move more info into the tooltip
so that I could pack as many keybindings as possible into the options view. At some point I realised that
you really need both a short and long description, so that you can keep keep the options view compact
but keep the keybindings menu easily searchable. For example 'Edit' is just 'Edit' in the options view but
it's 'Edit (start interactive rebase)' in the keybindings menu. It's pretty important to have the
'start interactive rebase' part there for users new to lazygit who don't know that the edit keybinding
is what we use for starting an interactive rebase.

And so, because it was so late in the game that I realised we really need a long and short description
(in addition to the tooltip), a lot of the long descriptions are quite short now. I don't actually mind
it because it makes it easier to skim the keybindings menu but it might be overboard in places.

You'll notice that some keybindings I deemed unworthy of the options view have longer descriptions,
because I could get away with it.